Added "bcryptjs"-module to crypt passwords.
Edited userLogin and addNewUser functions from userController.js to use bcryptjs.

addNewUser checks now if alias is already taken and prevents registration if it is.

Added frontend validation for alias in signupform.handlebars with jquery/ajax to getBooleanIfAliasInDB -function in userController.js to show user if alias is taken.

Some css changes in style.css